Some Changes in paths and spelling
The directory path on the used EV3 brick I am using was different than the one in the script. This change made it work for me. The code did not work at all without this change to all the sensor paths that needed the change made.
I changed it from:
ev3devices/in1/value0
to:
ev3devices/ev3-ports:in1/value0

I removed the extra "l" from all the instances of the word roll. Changed Rolll to Roll in all Variables. It just makes me happier. I don't think it will help the code work better.

I fixed spelling errors in the comments because of my CDO (that's OCD spelled in alphabetical order). I does not help the code work.